pcn-k8s: Reformat and simplify other packages
The networkpolicies and pcnfirewall packages have been simplified and
made compatible with the changes made previously on the controllers.
As a result, many imports and variables have been removed, and numerous
functions have been trimmed down (starting with their parameters).
