+++ src/components/k8s/pcn_k8s/controllers/controllers.go
@@ -23,7 +23,6 @@ const (
 )
 
 func init() {
-
 	logger = log.New()
 
 	// Only log the debug severity or above.
