+++ src/components/k8s/pcn_k8s/controllers/doc.go
@@ -2,10 +2,3 @@
 // They are used to watch for events about resources or to get
 // information about them.
 package controllers
-
-const (
-	// PC is the short name of the Pod Controller
-	PC string = "PodController"
-	// NC is the short name of the Namespace Controller
-	NC string = "NamespaceController"
-)
