+++ src/components/k8s/pcn_k8s/networkpolicies/doc.go
@@ -4,13 +4,3 @@
 // Then, the network policy manager, will take care of managing
 // firewall managers inside its node
 package networkpolicies
-
-const (
-	// KPP is the short name of the Kubernetes Network Policy Parser
-	KPP = "K8s Policy Parser"
-	// PM is the short name of the Network Policy Manager
-	PM = "Policy Manager"
-	// UnscheduleThreshold is the maximum number of minutes
-	// a firewall manager should live with no pods monitored.
-	UnscheduleThreshold = 5
-)
